Establish test suite
This commit establishes a test suite for the financial app. This does not create tests for all current methods; it only sets up a framework with one example. It also updates the README for instructions on running the tests.

In its current state, the tests assume that the user has an emulator running in a separate terminal and has set up the string variables for project, instance, and database IDs.

This PR is also just a start to setting up tests. More work will be done on improving the tests, but this provides a baseline.