Use IntegrationTestEnv (#28)
This commit introduces IntegrationTestEnv into this app's tests. This testing environment is used by both java-spanner and java-spanner-jdbc (and possibly other repos) to set up an integration test environment. This includes initializing a test database and deleting it at the end of tests. Tests can be ran with `mvn integration-test`.

We also get rid of the explicit assignment of variables for project, instance, and database ID, instead getting these variables from IntegrationTestEnv, which is influenced by both the environment in which the tests are run (including general gcloud setup) and the config in the Maven pom.xml file.

A lot of the decisions made in this commit were influenced by previous uses of IntegrationTestEnv and this pom.xml file.

Another part of the commit is the removal of parametrized tests:

ParameterizedTest tests are run before the BeforeClass methods. Before, I had used a method (suggested in a related Stack Overflow forum) to get around this by replacing my BeforeClass methods with a simple static {}. However, this had some strange interactions with the ClassRule (which was introduced with IntegrationTestEnv).

Switching the static back to BeforeClass to ensure the order of ClassRule to BeforeClass to Before to tests, I had to rework the code to no longer use ParameterizedTest, instead opting to set the SpannerDaoInterface implementation through environment.

By default, the tests use the Java client implementation, but adding `-DSPANNER_USE_JDBC=true` to `mvn integration test` switches to the JDBC implementation.