Establish basic workload generator (#38)
This commit establishes the basic foundation for a workload generator. This current workload generator first creates and then continuously sends `moveAccountBalance` requests for these accounts.

Notes on the current generator:
1) requires the server to be ran separately (with project, instance, and database set up)
    - requires the database to be created beforehand (`$ gcloud spanner databases create test-database --ddl-file src/main/java/com/google/finapp/schema.sdl`)
2) only uses `createAccount` and `moveAccountBalance`
3) single threaded

I have not touched any Java logging configuration. Running the Maven command as is above will cause Maven to output all of the log messages to console. This can be redirected to a log file by appending `-l <log name>` to the Maven command.