Make workload generator multithreaded (#50)
This commit restructures the workload generator to be multithreaded. This change makes it so that clients run indefinitely, creating their own accounts and performing methods at random with those accounts. This removes the need for the `seedData` method of `WorkloadMain` since the account ids do not need to be generated beforehand.